requirements.txt
Shapely==1.6.4.post2
pyexifinfo==0.4.0
Pillow==6.2.0
opencv-python-headless==4.1.2.30
Flask==1.1.1
